Moved itly audit checks to top-level to prevent logging to kits Added auditing for Kit events Reverted to Java 1.6
