+++ android-core/src/main/java/com/mparticle/internal/MessageManager.java
@@ -52,16 +52,12 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
-import ly.iterative.itly.Event;
-import ly.iterative.itly.android.Itly;
-
 /**
  * This class is primarily responsible for generating BaseMPMessage objects, and then adding them to a
  * queue which is then processed in a background thread for further processing and database storage.
  *
  */
 public class MessageManager implements MessageManagerCallbacks, ReportingManager {
-    private static final String ITLY_ATTRIBUTE_KEY = "$itly";
     private static Context sContext = null;
     static SharedPreferences sPreferences = null;
     static volatile boolean devicePerformanceMetricsDisabled;
@@ -70,7 +66,6 @@ public class MessageManager implements MessageManagerCallbacks, ReportingManager
     private ConfigManager mConfigManager = null;
     private MParticleDBManager mMParticleDBManager;
     private MParticle.OperatingSystem mOperatingSystem;
-    private Itly mItly;
 
     /**
      * These two threads are used to do the heavy lifting.
@@ -168,7 +163,6 @@ public class MessageManager implements MessageManagerCallbacks, ReportingManager
         mUploadHandler = new UploadHandler(options.getContext(), sUploadHandlerThread.getLooper(), configManager, appStateManager, this, dbManager);
         sPreferences = options.getContext().getSharedPreferences(Constants.PREFS_FILE, Context.MODE_PRIVATE);
         mInstallType = options.getInstallType();
-        mItly = options.getItly();
     }
 
     private static TelephonyManager getTelephonyManager() {
@@ -326,47 +320,8 @@ public class MessageManager implements MessageManagerCallbacks, ReportingManager
         }
     }
 
-    /**
-     * Logs event via Itly depending on configuration. Returns true if event was logged, false otherwise
-     *
-     * @param eventName
-     * @param eventType
-     * @param event
-     * @return
-     */
-    private boolean logEventToItly(String eventName, MParticle.EventType eventType, BaseEvent event) {
-        // Check for $itly property
-        if (event.getCustomAttributes() != null && event.getCustomAttributes().containsKey(ITLY_ATTRIBUTE_KEY)) {
-            // if $itly is found, remove it and track via standard mParticle logic
-            event.getCustomAttributes().remove(ITLY_ATTRIBUTE_KEY);
-        } else if (mItly != null) {
-            // Construct metadata
-            Map<String, Object> mpMetadata = new HashMap<>();
-            mpMetadata.put("eventType", eventType);
-            if (event.getCustomFlags() != null) {
-                mpMetadata.put("customFlags", event.getCustomFlags());
-            }
-            Map<String, Map<String, Object>> metadata = new HashMap<>();
-            metadata.put("mparticle", mpMetadata);
-
-            mItly.track(new Event(
-                eventName,
-                event.getCustomAttributes(),
-                null,
-                null,
-                metadata
-            ));
-            return true;
-        }
-        return false;
-    }
-
     public BaseMPMessage logEvent(MPEvent event, String currentActivity) {
         if (event != null) {
-            if (logEventToItly(event.getEventName(), event.getEventType(), event)) {
-                return null;
-            }
-
             try {
 
                 BaseMPMessage message = event.getMessage()
@@ -391,10 +346,6 @@ public class MessageManager implements MessageManagerCallbacks, ReportingManager
 
     public BaseMPMessage logEvent(CommerceEvent event) {
         if (event != null) {
-            if (logEventToItly(event.getEventName(), MParticle.EventType.Other, event)) {
-                return null;
-            }
-
             try {
                 BaseMPMessageBuilder builder = event.getMessage();
                 BaseMPMessage message = builder
